Qt6: qHash now returns size_t
git-svn-id: svn://scribus.net/trunk/Scribus@24871 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pdflib_core.cpp
@@ -1660,7 +1660,7 @@ PdfId PDFLibCore::PDF_EmbedFontObject(const QByteArray& font, const QByteArray& 
 
 QByteArray PDFLibCore::PDF_GenerateSubsetTag(const QByteArray& fontName, const QList<uint>& usedGlyphs)
 {
-	uint hash, mod;
+	size_t hash, mod;
 	QVector<uint> glyphVec = usedGlyphs.toVector();
 
 	hash = qHashBits(fontName.constData(), fontName.size(), 0);
